+++ Server/module/FcModule/FcService.js
@@ -1,12 +1,9 @@
 const bcrypt = require("bcrypt");
 const db = require("../../database");
 const FcModule = require("./FC");
-const geoip = require('geoip-lite');
-const {getIPAddress} = require('../../routers/Common');
-const os = require('os');
-
-
-
+const geoip = require("geoip-lite");
+const { getIPAddress } = require("../../routers/Common");
+const os = require("os");
 
 module.exports = {
   // get all entry
@@ -16,14 +13,15 @@ module.exports = {
       return res.status(200).json({
         status: 200,
         success: 1,
-        data:allData
-    });
+        msg:`data found..`,
+        data: allData,
+      });
     } catch (error) {
       return res.status(500).json({
         status: 500,
         msg: "Internal sarver error!!",
-        success: 0
-    });
+        success: 0,
+      });
     }
   },
 
@@ -32,20 +30,25 @@ module.exports = {
     try {
       const data = await FcModule.findByPk(req.params.id);
       if (!data)
-        res.status(200).json(`Data not found with fc id :${req.params.id}`);
+        return res.status(200).json({
+          status: 200,
+          msg: `data not found with id:${req.params.id}`,
+          success: 0,
+          data: data,
+        });
       else {
         return res.status(200).json({
           status: 200,
           success: 1,
-          data:data
-      });
+          data: data,
+        });
       }
     } catch (error) {
       return res.status(500).json({
         status: 500,
         msg: "Internal sarver error!!",
-        success: 0
-    });
+        success: 0,
+      });
     }
   },
 
@@ -57,16 +60,24 @@ module.exports = {
         (err, result) => {}
       );
       if (data[0].length == 0)
-        res
-          .status(404)
-          .json(`Data not found with fc email id : ${req.body.email}`);
-      else res.status(200).json(data[0][0]);
+        return res.status(200).json({
+          status: 200,
+          msg: `data not found with email:${req.params.email}`,
+          success: 0,
+        });
+      else
+        return res.status(200).json({
+          status: 200,
+          msg: `data found with email:${req.params.email}`,
+          success: 1,
+          data: data,
+        });
     } catch (error) {
       return res.status(500).json({
         status: 500,
         msg: "Internal sarver error!!",
-        success: 0
-    });
+        success: 0,
+      });
     }
   },
 
@@ -78,10 +89,9 @@ module.exports = {
       const salt = bcrypt.genSaltSync(10);
       const hash = bcrypt.hashSync(rawData.password, salt);
       rawData.password = hash;
-      rawData.createAt=new Date();
-      rawData.updateAt=new Date();
+      rawData.createAt = new Date();
+      rawData.updateAt = new Date();
       const data = await FcModule.create(rawData);
-
 
       const ipAddress = await getIPAddress();
       const geo = geoip.lookup(ipAddress);
@@ -91,21 +101,23 @@ module.exports = {
       const userEmail = data.dataValues.email;
       const type = "Fc";
 
-      await db.query(`INSERT INTO registerActivity (userEmail, userType, latitude, longitude, ipAddress, systemName)
-      VALUES ('${userEmail}', '${type}', '${latitude}','${longitude}', '${ipAddress}', '${systemName}');`, (err, result)=>{});
-    
+      await db.query(
+        `INSERT INTO registerActivity (userEmail, userType, latitude, longitude, ipAddress, systemName)
+      VALUES ('${userEmail}', '${type}', '${latitude}','${longitude}', '${ipAddress}', '${systemName}');`,
+        (err, result) => {}
+      );
 
       return res.status(201).json({
         status: 201,
         msg: "create new data successfully",
-        success: 1
-    });
+        success: 1,
+      });
     } catch (error) {
       return res.status(500).json({
         status: 500,
         msg: "Internal sarver error!!",
-        success: 0
-    });
+        success: 0,
+      });
     }
   },
 
@@ -114,43 +126,75 @@ module.exports = {
     try {
       const find = await FcModule.findByPk(req.params.id);
       if (!find)
-        res.status(200).json(`Data not found with fc id :${req.params.id}`);
+        return res.status(200).json({
+          status: 200,
+          msg: `data found with id:${req.params.id}`,
+          success: 0,
+        });
       else {
         const rawData = req.body;
-        rawData.updateAt=new Date();
+        rawData.updateAt = new Date();
         const data = await FcModule.update(rawData, {
           where: {
             fc_id: req.params.id,
           },
         });
-        if (data[0] == 0) res.status(200).json("already updated...");
-        else res.status(200).json("update successfully....");
+        if (data[0] == 0)
+          return res.status(200).json({
+            status: 200,
+            msg: `already updated...`,
+            success: 0,
+          });
+        else
+          return res.status(200).json({
+            status: 200,
+            msg: `updated successfully`,
+            success: 1,
+          });
       }
     } catch (error) {
       return res.status(500).json({
         status: 500,
         msg: "Internal sarver error!!",
-        success: 0
-    });
+        success: 0,
+      });
     }
   },
 
   // deleted entry by id
   deleteFcById: async (req, res) => {
     try {
-      const data = await FcModule.destroy({
-        where: {
-          fc_id: req.params.id,
-        },
-      });
-      if (data == 1) res.status(200).json("deleted successfully");
-      else res.status(200).json(`Data not found with fc id :${req.params.id}`);
+      const find = await FcModule.findByPk(req.params.id);
+      if (!find)
+        return res.status(200).json({
+          status: 200,
+          msg: `data found with id:${req.params.id}`,
+          success: 0,
+        });
+      else {
+        const data = await FcModule.destroy({
+          where: {
+            fc_id: req.params.id,
+          },
+        });
+        if (data == 1) return res.status(200).json({
+          status: 200,
+          msg: `deleted successfully`,
+          success: 1,
+        });
+        else
+        return res.status(200).json({
+          status: 200,
+          msg: `something wrong...`,
+          success: 0,
+        });
+      }
     } catch (error) {
       return res.status(500).json({
         status: 500,
         msg: "Internal sarver error!!",
-        success: 0
-    });
+        success: 0,
+      });
     }
   },
 };
